# api-docs/swagger.yaml
openapi: 3.0.0
paths:
  /v1/expenses:
    post:
      operationId: ExpensesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpensesCreateDTO'
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expenses'
        '409':
          description: Conflict - already exists
      summary: Create a new expense
      tags: &ref_0
        - expenses
    get:
      operationId: ExpensesController_list
      parameters:
        - name: queryFilter
          required: false
          in: query
          description: Search term across multiple fields
          schema:
            type: string
        - name: fromDate
          required: false
          in: query
          description: Filter by start date
          schema:
            format: date-time
            type: string
        - name: toDate
          required: false
          in: query
          description: Filter by end date
          schema:
            format: date-time
            type: string
        - name: pagination
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/PaginationDTO'
        - required: false
          description: Search term across multiple fields
          name: queryFilter
          in: query
          schema:
            type: string
        - required: false
          description: Filter by start date
          name: fromDate
          in: query
          schema:
            format: date-time
            type: string
        - required: false
          description: Filter by end date
          name: toDate
          in: query
          schema:
            format: date-time
            type: string
        - name: pagination
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/PaginationDTO'
      responses:
        '200':
          description: Paginated list of expenses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResult'
      summary: List expenses with pagination and filtering
      tags: *ref_0
  /v1/expenses/aggregations:
    get:
      operationId: ExpensesController_getAggregations
      parameters:
        - name: clientId
          required: false
          in: query
          description: Filter by client ID
          schema:
            type: string
        - name: category
          required: false
          in: query
          description: Filter by category
          schema:
            type: string
        - name: currency
          required: false
          in: query
          description: Filter by currency
          schema:
            example: USD
            type: string
        - name: year
          required: false
          in: query
          description: Filter by year
          schema:
            example: 2026
            type: number
        - name: month
          required: false
          in: query
          description: Filter by month (1-12)
          schema:
            example: 6
            type: number
        - name: day
          required: false
          in: query
          description: Filter by day
          schema:
            example: 3
            type: number
        - required: false
          description: Filter by client ID
          name: clientId
          in: query
          schema:
            type: string
        - required: false
          description: Filter by category
          name: category
          in: query
          schema:
            type: string
        - required: false
          description: Filter by currency
          name: currency
          in: query
          schema:
            example: USD
            type: string
        - required: false
          description: Filter by year
          name: year
          in: query
          schema:
            example: 2026
            type: number
        - required: false
          description: Filter by month (1-12)
          name: month
          in: query
          schema:
            example: 6
            type: number
        - required: false
          description: Filter by day
          name: day
          in: query
          schema:
            example: 3
            type: number
      responses:
        '200':
          description: List of expense aggregations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpensesAggregation'
      summary: Get expense aggregations by filters
      tags: *ref_0
  /v1/expenses/{id}:
    get:
      operationId: ExpensesController_getOne
      parameters:
        - name: id
          required: true
          in: path
          description: Expense UUID
          schema:
            type: string
      responses:
        '200':
          description: Expense details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expenses'
        '404':
          description: Expense not found
      summary: Get expense by ID
      tags: *ref_0
  /v1/expenses/{id}/status:
    put:
      operationId: ExpensesController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          description: Expense UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpensesUpdateStatusDTO'
      responses:
        '200':
          description: Expense updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expenses'
        '404':
          description: Expense not found or invalid status
      summary: Update expense status
      tags: *ref_0
info:
  title: Expenses API
  description: Expenses management microservice
  version: '1.0'
  contact: {}
tags:
  - name: expenses
    description: ''
servers: []
components:
  schemas:
    ExpensesCreateDTO:
      type: object
      properties:
        clientId:
          type: string
          description: Client identifier (UUID, MongoID, or other string identifier)
        amount:
          type: number
          description: Expense amount
          example: 100.5
        currency:
          type: string
          description: ISO 4217 currency code
          example: USD
        category:
          type: string
          description: Expense category
          example: Food
        description:
          type: string
          description: Expense description
          example: Lunch at restaurant
        status:
          type: string
          description: Expense status
          enum:
            - pending
            - reviewed
            - voided
          default: pending
      required:
        - clientId
        - amount
        - currency
        - category
        - description
        - status
    Expenses:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier (UUID)
        amount:
          type: number
          description: Expense amount
          example: 100.5
        currency:
          type: string
          description: ISO 4217 currency code
          example: USD
        category:
          type: string
          description: Expense category
          example: Food
        description:
          type: string
          description: Expense description
          example: Lunch at restaurant
        clientId:
          type: string
          description: Client identifier
          example: uuid-or-mongoid
        status:
          type: string
          description: Expense status
          enum:
            - pending
            - reviewed
            - voided
        creationDate:
          format: date-time
          type: string
          description: Creation timestamp
        year:
          type: number
          description: Year of expense
          example: 2026
        month:
          type: number
          description: Month of expense (1-12)
          example: 6
        day:
          type: number
          description: Day of month
          example: 3
      required:
        - id
        - amount
        - currency
        - category
        - description
        - clientId
        - status
        - creationDate
        - year
        - month
        - day
    PaginationDTO:
      type: object
      properties:
        page:
          type: number
          description: Page number
          example: 1
          default: 1
        limit:
          type: number
          description: Items per page
          example: 25
          default: 25
        sort:
          type: string
          description: Sort order
          enum:
            - asc
            - desc
          default: asc
        countTotal:
          type: boolean
          description: Whether to count total items
          default: true
      required:
        - page
        - limit
        - sort
        - countTotal
    PaginationResult:
      type: object
      properties:
        data:
          description: List of items
          type: array
          items:
            type: array
        page:
          type: number
          description: Current page number
        limit:
          type: number
          description: Items per page
        sort:
          type: string
          description: Sort order
          enum:
            - asc
            - desc
        totalItems:
          type: number
          description: Total number of items
      required:
        - data
        - page
        - limit
        - sort
        - totalItems
    ExpensesAggregation:
      type: object
      properties:
        clientId:
          type: string
          description: Client identifier
        category:
          type: string
          description: Expense category
        year:
          type: number
          description: Year of expense
        month:
          type: number
          description: Month of expense (1-12)
        day:
          type: number
          description: Day of month
        currency:
          type: string
          description: ISO 4217 currency code
        count:
          type: number
          description: Number of expenses in this group
        totalAmount:
          type: number
          description: Total amount of expenses in this group
        creationDate:
          format: date-time
          type: string
          description: Creation timestamp
        updatedAt:
          format: date-time
          type: string
          description: Last update timestamp
      required:
        - clientId
        - category
        - year
        - month
        - day
        - currency
        - count
        - totalAmount
        - creationDate
        - updatedAt
    ExpensesUpdateStatusDTO:
      type: object
      properties:
        status:
          type: string
          description: Expense status
          enum:
            - pending
            - reviewed
            - voided
          default: pending
      required:
        - status